Separate csv formatting files and components
and remove todo
